+++ src/main/java/lk/ijse/gdse/service/util/VehicleServiceIMPL.java
@@ -3,6 +3,7 @@ package lk.ijse.gdse.service.util;
 
 import lk.ijse.gdse.dto.VehicleDTO;
 import lk.ijse.gdse.entity.VehicleEntity;
+import lk.ijse.gdse.exception.NotFoundException;
 import lk.ijse.gdse.repo.VehicleRepo;
 import lk.ijse.gdse.service.VehicleService;
 import lk.ijse.gdse.util.Converter;
@@ -21,6 +22,7 @@ public class VehicleServiceIMPL implements VehicleService {
     private final Converter convert;
 
     private final VehicleRepo vehicleRepo;
+
     @Override
     public VehicleDTO saveVehicle(VehicleDTO vehicleDTO) {
         return convert.toVehicleDTO(vehicleRepo.save(convert.toVehicleEntity(vehicleDTO)));
@@ -28,26 +30,39 @@ public class VehicleServiceIMPL implements VehicleService {
 
     @Override
     public VehicleDTO getSelectedVehicle(String vehicle_id) {
-        return convert.toVehicleDTO(vehicleRepo.findById(vehicle_id).get());
+        Optional<VehicleEntity> byId = vehicleRepo.findById(vehicle_id);
+        if (!byId.isPresent()) {
+            throw new NotFoundException("The vehicle id cannot be found  :" + vehicle_id);
+        }
+        return convert.toVehicleDTO(byId.get());
+        // vehicleRepo.findById(vehicle_id).orElseThrow(()->new NotFoundException("Vehicle Id Not Found :"+ vehicle_id));
+
     }
 
     @Override
-    public void updateVehicle(VehicleDTO vehicleDTO) {
-        Optional<VehicleEntity> vehicleEntity=vehicleRepo.findById(vehicleDTO.getVehicle_id());
-        if (vehicleEntity.isPresent()){
-            vehicleEntity.get().setDriver_name(vehicleDTO.getDriver_name());
-            vehicleEntity.get().setVehicle_brand(vehicleDTO.getVehicle_brand());
-            vehicleEntity.get().setVehicle_category(vehicleDTO.getVehicle_category());
-            vehicleEntity.get().setFuel_type(vehicleDTO.getFuel_type());
-            vehicleEntity.get().setFuel_usage(vehicleDTO.getFuel_usage());
-            vehicleEntity.get().setSeat_capacity(vehicleDTO.getSeat_capacity());
-            vehicleEntity.get().setTransmission(vehicleDTO.getTransmission());
+    public void updateVehicle(String vehicle_id, VehicleDTO vehicleDTO) {
+        Optional<VehicleEntity> vehicleEntity = vehicleRepo.findById(vehicle_id);
+        if (!vehicleEntity.isPresent()) {
+            throw new NotFoundException("The vehicle id cannot be found  :" + vehicle_id);
+
         }
+        VehicleEntity vehicle = vehicleEntity.get();
+        vehicle.setDriver_name(vehicleDTO.getDriver_name());
+        vehicle.setVehicle_brand(vehicleDTO.getVehicle_brand());
+        vehicle.setVehicle_category(vehicleDTO.getVehicle_category());
+        vehicle.setFuel_type(vehicleDTO.getFuel_type());
+        vehicle.setFuel_usage(vehicleDTO.getFuel_usage());
+        vehicle.setSeat_capacity(vehicleDTO.getSeat_capacity());
+        vehicle.setTransmission(vehicleDTO.getTransmission());
 
     }
 
     @Override
     public void deleteVehicle(String vehicle_id) {
-      vehicleRepo.deleteById(vehicle_id);
+        Optional<VehicleEntity> byId = vehicleRepo.findById(vehicle_id);
+        if (!byId.isEmpty()) {
+            throw new NotFoundException("The vehicle id cannot be found  :" + vehicle_id);
+        }
+        vehicleRepo.deleteById(vehicle_id);
     }
 }
